+++ app/(tabs)/(categories)/categories.tsx
@@ -3,40 +3,44 @@ import { IconSymbol } from '@/components/ui/icon-symbol';
 import { className } from '@/constants/classNames';
 import colors from '@/constants/nativewindColors';
 import { useRouter } from 'expo-router';
-import React, { useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import { Pressable, Text, TouchableOpacity, useColorScheme, View } from 'react-native';
 import DraggableFlatList from "react-native-draggable-flatlist";
 import { SafeAreaView, useSafeAreaInsets } from 'react-native-safe-area-context';
 
 import CustomHeader from '@/components/elements/custom-header';
 import { useCategories } from '@/context/category-context';
+import { isDeepEqual } from '@/functions/handling';
 import { Category } from '@/types/models';
 
 const CategoriesTab = () => {
   const router = useRouter();
-  const colorScheme = useColorScheme(); // this is only necessary since I need this variable for the IconSymbol
-  const isDarkMode = colorScheme === 'dark';
+  const colorSchemeKey = useColorScheme() === 'dark' ? 'dark' : 'light';
   const insets = useSafeAreaInsets(); // used so that I can use manual extra padding to the flatlist so that it does not hide under the tabbar
 
+  // the globally used categories array
   const { categories, setCategories, addCategory, removeCategory } = useCategories();
+  // the categories used on this screen, save globally only upon clicking "Save"
+  const [localCategories, setLocalCategories] = useState<Category[]>(categories);
   const [editMode, setEditMode] = useState(false);
   const [addCategoryModalVisible, setAddCategoryModalVisible] = useState(false);
+  const hasChanges = useMemo(() => !isDeepEqual(localCategories, categories), [localCategories, categories]);
 
   const RenderBox = editMode ? View : TouchableOpacity;
   const renderItem = ({ item, drag, isActive }: { item: Category, drag : any, isActive : any }) => (
     <RenderBox
       className={className.item}
-      style={{ backgroundColor: item.color ?? (isDarkMode ? colors.dark.content.accent : colors.light.content.accent) }}
+      style={{ backgroundColor: item.color ?? colors[colorSchemeKey].content.accent }}
       onPress={() => router.push(`/category/${item.id}`)} // if RenderBox is a View, this prop is simply ignored
     >
       <Text className={className.text.item}>{item.name}</Text>
       {editMode && (
         <View className="flex-row">
           <Pressable onLongPress={drag} delayLongPress={0} className={className.button.icon}>
-        <IconSymbol size={28} name="reorder" color={isActive ? (isDarkMode ? colors.dark.content.tertiary : colors.light.content.tertiary) : (isDarkMode ? colors.dark.surface.sunken : colors.light.surface.sunken)} />
+        <IconSymbol size={28} name="reorder" color={isActive ? colors[colorSchemeKey].content.tertiary : colors[colorSchemeKey].surface.sunken} />
         </Pressable>
         <TouchableOpacity onPress={() => removeCategory(item.id)} className={className.button.icon}>
-          <IconSymbol size={28} name="delete" color={isDarkMode ? colors.dark.content.negative : colors.light.content.negative} />
+          <IconSymbol size={28} name="delete" color={ colors[colorSchemeKey].content.negative} />
           </TouchableOpacity>
         </View>
       )}
@@ -53,17 +57,26 @@ const CategoriesTab = () => {
             {"+"}
           </Text>
         </TouchableOpacity>
-        <TouchableOpacity className={className.button.secondary} 
-        onPress={() => setEditMode(!editMode)}>
+        <TouchableOpacity 
+          className={
+            hasChanges
+            ? className.button.primary
+            : className.button.secondary
+          } 
+          onPress={() => {
+            if(hasChanges) setCategories(localCategories);
+            setEditMode(!editMode);
+          }
+        }>
           <Text className="px-5 text-xl ">
-            {editMode ? <IconSymbol size={28} name="done" color={ isDarkMode ? colors.dark.content.primary : colors.light.content.primary}/> : <IconSymbol size={28} name="edit" color={ isDarkMode ? colors.dark.content.primary : colors.light.content.primary} />}
+            {editMode ? <IconSymbol size={28} name="done" color={ hasChanges ? colors[colorSchemeKey].content.onAccent : colors[colorSchemeKey].content.primary}/> : <IconSymbol size={28} name="edit" color={ colors[colorSchemeKey].content.primary } />}
           </Text>
         </TouchableOpacity>
       </CustomHeader>
 
       <DraggableFlatList
-        data={categories}
-        onDragEnd={({ data }) => setCategories(data)}
+        data={localCategories}
+        onDragEnd={({ data }) => setLocalCategories(data)}
         keyExtractor={item => item.id}
         renderItem={renderItem}
         className="rounded-2xl"